% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_accuracy_measure.R
\name{bonsai_accuracy_measure}
\alias{bonsai_accuracy_measure}
\title{Chosen accuracy measure}
\usage{
bonsai_accuracy_measure(y, error, error_function)
}
\arguments{
\item{y}{A time series}

\item{error}{the error created by some cross-validation method}
}
\value{
A ? with all of the forecast means
}
\description{
Calculates the chosen accuracy measure
}
