% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_calculate_errors.R
\name{bonsai_calculate_errors}
\alias{bonsai_calculate_errors}
\title{Creates a list with all specified errors measures}
\usage{
bonsai_calculate_errors(y, list_errors, error_function)
}
\arguments{
\item{y}{A time series to be forecasted}

\item{list_errors}{A list of errors to apply}

\item{error_function}{a function with all errors measures}
}
\value{
A list with all of the calculated error measures
}
\description{
Takes in a list of errors and return a list of error measuares
}
