% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_cv.R
\name{bonsai_cv}
\alias{bonsai_cv}
\title{Time series cross-validation with extra parameters}
\usage{
bonsai_cv(y, forecast_function, h = 1, window = NULL, Start = 1,
  max_fold = NULL, min_lenght = 0, ...)
}
\arguments{
\item{y}{Univariate time series}

\item{forecast_function}{Function to return an object of class forecast. Its first argument must be a univariate time series, and it must have an argument h for the forecast horizon.}

\item{h}{Forecast horizon}

\item{window}{Length of the rolling window, if NULL, a rolling window will not be used.}

\item{Start}{how many observations before starting to calculate errors}

\item{max_fold}{how many times should the function calculate errors}

\item{min_lenght}{how many observations are nescessary before starting to calculate errors}
}
\value{
Numerical time series object containing the forecast errors as a vector (if h=1) and a matrix otherwise. The time index corresponds to the last period of the training data. The columns correspond to the forecast horizons.

A matrix with the weighted object
}
\description{
bonsai_cv computes the forecast errors obtained by applying forecast_function to subsets of the time series y using a diverse set of rules.
}
\details{
Let y contain the time series y[1:T]. Then forecastfunction is applied successively to the time series y[1:t], for t=1,…,T-h, making predictions f[t+h]. The errors are given by e[t+h] = y[t+h]-f[t+h]. If h=1, these are returned as a vector, e[1:T]. For h>1, they are returned as a matrix with the hth column containing errors for forecast horizon h. The first few errors may be missing as it may not be possible to apply forecastfunction to very short time series.
}
\author{
Rob Hyndman created the tsCV {forecast}, heavily based on it
}
