% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_cv_models.R
\name{bonsai_cv_models}
\alias{bonsai_cv_models}
\title{Uses cross validation with a list of functions}
\usage{
bonsai_cv_models(y, list_functions, h, window = NULL, Start = 1,
  min_lenght = 0, max_fold = NULL)
}
\arguments{
\item{y}{Univariate time series}

\item{list_functions}{a list of functions to test}

\item{h}{Forecast horizon}

\item{window}{Length of the rolling window, if NULL, a rolling window will not be used.}

\item{Start}{how many observations before starting to calculate errors}

\item{min_lenght}{how many observations are nescessary before starting to calculate errors}

\item{max_fold}{how many times should the function calculate errors}
}
\value{
Numerical time series object containing the forecast errors as a vector (if h=1) and a matrix otherwise. The time index corresponds to the last period of the training data. The columns correspond to the forecast horizons.

A matrix with the weighted object
}
\description{
Commodity function to facilitate the usage of the bonsai.cv function, may be depriciated
}
\details{
Let y contain the time series y[1:T]. Then forecastfunction is applied successively to the time series y[1:t], for t=1,…,T-h, making predictions f[t+h]. The errors are given by e[t+h] = y[t+h]-f[t+h]. If h=1, these are returned as a vector, e[1:T]. For h>1, they are returned as a matrix with the hth column containing errors for forecast horizon h. The first few errors may be missing as it may not be possible to apply forecastfunction to very short time series.
}
\author{
Rob Hyndman created the tsCV {forecast}, heavily based on it
}
