% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_error_matrix.R
\name{bonsai_error_matrix}
\alias{bonsai_error_matrix}
\title{Create error matrix}
\usage{
bonsai_error_matrix(list_accuracy)
}
\arguments{
\item{list_accuracy}{a list with all the accuracy measures}
}
\value{
A list with all the errors
}
\description{
Creates the error matrix instead of an weight matrix, usefull for some methods
}
