% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_forecast_extractor.R
\name{bonsai_forecast_extractor}
\alias{bonsai_forecast_extractor}
\title{Extracts specifc parts of forecasts}
\usage{
bonsai_forecast_extractor(list_forecasts, extract_object = "mean",
  forecast_functions, colname_logic = "h=")
}
\arguments{
\item{list_forecasts}{a list of forecasts generated by bonsai_forecast_saver}

\item{extract_object}{what to extract usually means, residuals, upper or lower}

\item{forecast_functions}{a commodity to name the rownames of thematrix may be depreciated}

\item{colname_logic}{a commodity to name the colnames of the matrix may be depreciated}
}
\value{
A matrix with the desired object
}
\description{
Takes in a list of forecasts generated by bonsai_forecast_saver and return a matrix with residuals or means or any other parameter
}
