% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_forecast_saver.R
\name{bonsai_forecast_saver}
\alias{bonsai_forecast_saver}
\title{Executes all forecasts in a list of functions}
\usage{
bonsai_forecast_saver(y, list_functions, h)
}
\arguments{
\item{y}{A time series to be forecasted}

\item{list_functions}{A list of functions to apply}

\item{h}{the horizon to forecast}
}
\value{
A list with all of the forecast means
}
\description{
Takes in any value and squares
}
