% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_invert_accuracy.R
\name{bonsai_invert_accuracy}
\alias{bonsai_invert_accuracy}
\title{Inverts a list of accuracy measures}
\usage{
bonsai_invert_accuracy(accuracy, if_zero = 100000L)
}
\arguments{
\item{accuracy}{A list of accuracy measures}

\item{if_zero}{What value to assingn when a perfect forecast happens}
}
\value{
A ? with all inverted accuracy`s
}
\description{
Inverts a list of accuracy measures to estimate weights, also deals with 0s
}
