% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_islands.R
\name{bonsai_islands}
\alias{bonsai_islands}
\title{Applis the islands method}
\usage{
bonsai_islands(matrix_errors, function_ordered_vector)
}
\arguments{
\item{matrix_errors}{A matrix of weights}

\item{function_ordered_vector}{a function to apply in the ordered vector}

\item{list_accuracy}{a list with all the accuracy measures}
}
\value{
A list with all the errors
}
\description{
chooses the best forecasts based on the Islands method
}
