% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_rank_maker.R
\name{bonsai_rank_maker}
\alias{bonsai_rank_maker}
\title{Calculate the ranks of all models and choose the selected ones}
\usage{
bonsai_rank_maker(matrix_weights, position_winner)
}
\arguments{
\item{matrix_weights}{A matrix of weights}

\item{position_winner}{The forecast model to be choosen}
}
\value{
A list with all of the weights
}
\description{
Calculates the position in a rank and assings 0 weight to all the losers.
}
