% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_results.R
\name{bonsai_results}
\alias{bonsai_results}
\title{Calculates weighted object}
\usage{
bonsai_results(matrix_weights, object_to_weight)
}
\arguments{
\item{matrix_weights}{a matrix of weights}

\item{object_to_weight}{the object that will get averaged weighted}
}
\value{
A matrix with the weighted object
}
\description{
uses the weights generated beforehand to calculte a new object, usually used for weighted forecasts or residuals
}
