% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_results_errors.R
\name{bonsai_results_errors}
\alias{bonsai_results_errors}
\title{Calculates the mase and smape errors}
\usage{
bonsai_results_errors(forecasts, y, test_ts)
}
\arguments{
\item{matrix_weights}{a matrix of weights}

\item{object_to_weight}{the object that will get averaged weighted}
}
\value{
A matrix with the weighted object
}
\description{
uses the weighted forecast generated beforehand to calculate error measures
}
