% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_test_greater_function.R
\name{bonsai_test_greater_function}
\alias{bonsai_test_greater_function}
\title{Calculate the weights of all models and choose the selected ones based on a function}
\usage{
bonsai_test_greater_function(matrix_weights, function_vector,
  if_all_bellow = FALSE, ...)
}
\arguments{
\item{matrix_weights}{A matrix of weights}

\item{function_vector}{a function to apply in the vector of weights}

\item{if_all_bellow}{a choice to avoid the case where all weights = 0}

\item{...}{arguments to be passed to the function}
}
\value{
A list with all of the weights
}
\description{
Checks all the weights using a function to determine a value usually mean.
}
