% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_test_greater_value.R
\name{bonsai_test_greater_value}
\alias{bonsai_test_greater_value}
\title{Calculate the weights of all models and choose the selected ones}
\usage{
bonsai_test_greater_value(matrix_weights, min_value,
  if_all_bellow = FALSE)
}
\arguments{
\item{matrix_weights}{A matrix of weights}

\item{min_value}{a minimun weight to use}

\item{if_all_bellow}{a choice to avoid the case where all weights = 0}
}
\value{
A list with all of the weights
}
\description{
Checks all the weights for values greather than min_value.
}
