% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_test_ranks.R
\name{bonsai_test_ranks}
\alias{bonsai_test_ranks}
\title{Tests all the ranks}
\usage{
bonsai_test_ranks(matrix_weights)
}
\arguments{
\item{matrix_weights}{A matrix of weights}
}
\value{
A list with all of the weights
}
\description{
Uses the rank maker function to try all possible combinations.
}
