% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_weight_matrix.R
\name{bonsai_weight_matrix}
\alias{bonsai_weight_matrix}
\title{Create weight matrix}
\usage{
bonsai_weight_matrix(list_accuracy)
}
\arguments{
\item{list_accuracy}{a list with all the accuracy measures}
}
\value{
A list with all the weights
}
\description{
Creates the weight matrix responsible for all the decisions made by the bonsain
}
