% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_weight_decision.R
\name{bonsai_weights_differentiate}
\alias{bonsai_weights_differentiate}
\title{Calculates all the "surviving"  weights}
\usage{
bonsai_weights_differentiate(matrix_weights)
}
\arguments{
\item{matrix_weights}{A matrix of weights}
}
\value{
A list with all the adjusted weights
}
\description{
Turn all non-zero weigths into the an weighted average
}
