% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_error_functions.R
\name{CV_Mean_Percentage_Error}
\alias{CV_Mean_Percentage_Error}
\title{Error functions}
\usage{
CV_Mean_Percentage_Error(y, error)
}
\arguments{
\item{y}{A time series}

\item{error}{the error created by some cross-validation method}

\item{h}{the horizon to forecast}
}
\value{
A list with all of the forecast means
}
\description{
Functions to be used in conjunction with bonsai_calculate_errors
}
