% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonsai_error_functions.R
\name{Root_Mean_Squared_Error}
\alias{Root_Mean_Squared_Error}
\title{Error functions}
\usage{
Root_Mean_Squared_Error(y = NULL, error)
}
\arguments{
\item{y}{A time series}

\item{error}{the error created by some cross-validation method}

\item{h}{the horizon to forecast}
}
\value{
A list with all of the forecast means
}
\description{
Functions to be used in conjunction with bonsai_calculate_errors
}
